+++ x/stake/keeper/delegation.go
@@ -5,18 +5,18 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/stake/types"
 )
 
-// load a delegator bond
+// load a delegation
 func (k Keeper) GetDelegation(ctx sdk.Context,
-	delegatorAddr, validatorAddr sdk.Address) (bond types.Delegation, found bool) {
+	delegatorAddr, validatorAddr sdk.Address) (delegation types.Delegation, found bool) {
 
 	store := ctx.KVStore(k.storeKey)
 	delegatorBytes := store.Get(GetDelegationKey(delegatorAddr, validatorAddr, k.cdc))
 	if delegatorBytes == nil {
-		return bond, false
+		return delegation, false
 	}
 
-	k.cdc.MustUnmarshalBinary(delegatorBytes, &bond)
-	return bond, true
+	k.cdc.MustUnmarshalBinary(delegatorBytes, &delegation)
+	return delegation, true
 }
 
 // load all delegations used during genesis dump
@@ -39,13 +39,15 @@ func (k PrivlegedKeeper) GetAllDelegations(ctx sdk.Context) (delegations []types
 	return delegations[:i] // trim
 }
 
-// load all bonds of a delegator
-func (k Keeper) GetDelegations(ctx sdk.Context, delegator sdk.Address, maxRetrieve int16) (bonds []types.Delegation) {
+// load all delegations for a delegator
+func (k Keeper) GetDelegations(ctx sdk.Context, delegator sdk.Address,
+	maxRetrieve int16) (delegations []types.Delegation) {
+
 	store := ctx.KVStore(k.storeKey)
 	delegatorPrefixKey := GetDelegationsKey(delegator, k.cdc)
 	iterator := sdk.KVStorePrefixIterator(store, delegatorPrefixKey) //smallest to largest
 
-	bonds = make([]types.Delegation, maxRetrieve)
+	delegations = make([]types.Delegation, maxRetrieve)
 	i := 0
 	for ; ; i++ {
 		if !iterator.Valid() || i > int(maxRetrieve-1) {
@@ -53,35 +55,35 @@ func (k Keeper) GetDelegations(ctx sdk.Context, delegator sdk.Address, maxRetrie
 			break
 		}
 		bondBytes := iterator.Value()
-		var bond types.Delegation
-		k.cdc.MustUnmarshalBinary(bondBytes, &bond)
-		bonds[i] = bond
+		var delegation types.Delegation
+		k.cdc.MustUnmarshalBinary(bondBytes, &delegation)
+		delegations[i] = delegation
 		iterator.Next()
 	}
-	return bonds[:i] // trim
+	return delegations[:i] // trim
 }
 
 // set the delegation
-func (k PrivlegedKeeper) SetDelegation(ctx sdk.Context, bond types.Delegation) {
+func (k PrivlegedKeeper) SetDelegation(ctx sdk.Context, delegation types.Delegation) {
 	store := ctx.KVStore(k.storeKey)
-	b := k.cdc.MustMarshalBinary(bond)
-	store.Set(GetDelegationKey(bond.DelegatorAddr, bond.ValidatorAddr, k.cdc), b)
+	b := k.cdc.MustMarshalBinary(delegation)
+	store.Set(GetDelegationKey(delegation.DelegatorAddr, delegation.ValidatorAddr, k.cdc), b)
 }
 
 // remove the delegation
-func (k PrivlegedKeeper) RemoveDelegation(ctx sdk.Context, bond types.Delegation) {
+func (k PrivlegedKeeper) RemoveDelegation(ctx sdk.Context, delegation types.Delegation) {
 	store := ctx.KVStore(k.storeKey)
-	store.Delete(GetDelegationKey(bond.DelegatorAddr, bond.ValidatorAddr, k.cdc))
+	store.Delete(GetDelegationKey(delegation.DelegatorAddr, delegation.ValidatorAddr, k.cdc))
 }
 
 // common functionality between handlers
 func (k PrivlegedKeeper) Delegate(ctx sdk.Context, delegatorAddr sdk.Address,
 	bondAmt sdk.Coin, validator types.Validator) (sdk.Tags, sdk.Error) {
 
-	// Get or create the delegator bond
-	bond, found := k.GetDelegation(ctx, delegatorAddr, validator.Owner)
+	// Get or create the delegator delegation
+	delegation, found := k.GetDelegation(ctx, delegatorAddr, validator.Owner)
 	if !found {
-		bond = types.Delegation{
+		delegation = types.Delegation{
 			DelegatorAddr: delegatorAddr,
 			ValidatorAddr: validator.Owner,
 			Shares:        sdk.ZeroRat(),
@@ -90,18 +92,18 @@ func (k PrivlegedKeeper) Delegate(ctx sdk.Context, delegatorAddr sdk.Address,
 
 	// Account new shares, save
 	pool := k.GetPool(ctx)
-	_, _, err := k.coinKeeper.SubtractCoins(ctx, bond.DelegatorAddr, sdk.Coins{bondAmt})
+	_, _, err := k.coinKeeper.SubtractCoins(ctx, delegation.DelegatorAddr, sdk.Coins{bondAmt})
 	if err != nil {
 		return nil, err
 	}
 	validator, pool, newShares := validator.AddTokensFromDel(pool, bondAmt.Amount)
-	bond.Shares = bond.Shares.Add(newShares)
+	delegation.Shares = delegation.Shares.Add(newShares)
 
-	// Update bond height
-	bond.Height = ctx.BlockHeight()
+	// Update delegation height
+	delegation.Height = ctx.BlockHeight()
 
 	k.SetPool(ctx, pool)
-	k.SetDelegation(ctx, bond)
+	k.SetDelegation(ctx, delegation)
 	k.UpdateValidator(ctx, validator)
 	tags := sdk.NewTags(
 		"action", []byte("delegate"),
@@ -128,6 +130,21 @@ func (k Keeper) GetUnbondingDelegation(ctx sdk.Context,
 	return ubd, true
 }
 
+// load a unbonding delegation and the associated delegation
+func (k Keeper) GetUnbondingDelegationDel(ctx sdk.Context, DelegatorAddr,
+	ValidatorAddr sdk.Address) (ubd types.UnbondingDelegation, delegation types.Delegation, found bool) {
+
+	ubd, found = k.GetUnbondingDelegation(ctx, DelegatorAddr, ValidatorAddr)
+	if !found {
+		return ubd, delegation, false
+	}
+	delegation, found = k.GetDelegation(ctx, ubd.DelegatorAddr, ubd.ValidatorAddr)
+	if !found {
+		panic("found unbonding delegation but not delegation object")
+	}
+	return ubd, delegation, true
+}
+
 // set the unbonding delegation and associated index
 func (k PrivlegedKeeper) SetUnbondingDelegation(ctx sdk.Context, ubd types.UnbondingDelegation) {
 	store := ctx.KVStore(k.storeKey)
@@ -135,6 +152,14 @@ func (k PrivlegedKeeper) SetUnbondingDelegation(ctx sdk.Context, ubd types.Unbon
 	ubdKey := GetUBDKey(ubd.DelegatorAddr, ubd.ValidatorAddr, k.cdc)
 	store.Set(ubdKey, bz)
 	store.Set(GetUBDByValIndexKey(ubd.DelegatorAddr, ubd.ValidatorAddr, k.cdc), ubdKey)
+}
+
+// remove the unbonding delegation object and associated index
+func (k PrivlegedKeeper) RemoveUnbondingDelegation(ctx sdk.Context, ubd types.UnbondingDelegation) {
+	store := ctx.KVStore(k.storeKey)
+	ubdKey := GetUBDKey(ubd.DelegatorAddr, ubd.ValidatorAddr, k.cdc)
+	store.Delete(ubdKey)
+	store.Delete(GetUBDByValIndexKey(ubd.DelegatorAddr, ubd.ValidatorAddr, k.cdc))
 }
 
 //_____________________________________________________________________________________
@@ -154,6 +179,25 @@ func (k Keeper) GetRedelegation(ctx sdk.Context,
 	return red, true
 }
 
+// load a unbonding delegation and the associated delegation
+func (k Keeper) GetRedelegationDel(ctx sdk.Context, DelegatorAddr, ValidatorSrcAddr,
+	ValidatorDstAddr sdk.Address) (red types.Redelegation, srcDelegation, dstDelegation types.Delegation, found bool) {
+
+	red, found = k.GetRedelegation(ctx, DelegatorAddr, ValidatorSrcAddr, ValidatorDstAddr)
+	if !found {
+		return red, srcDelegation, dstDelegation, false
+	}
+	srcDelegation, found = k.GetDelegation(ctx, red.DelegatorAddr, red.ValidatorSrcAddr)
+	if !found {
+		panic("found redelegation but not source delegation object")
+	}
+	dstDelegation, found = k.GetDelegation(ctx, red.DelegatorAddr, red.ValidatorDstAddr)
+	if !found {
+		panic("found redelegation but not source delegation object")
+	}
+	return red, srcDelegation, dstDelegation, true
+}
+
 // set a redelegation and associated index
 func (k PrivlegedKeeper) SetRedelegation(ctx sdk.Context, red types.Redelegation) {
 	store := ctx.KVStore(k.storeKey)
@@ -162,4 +206,13 @@ func (k PrivlegedKeeper) SetRedelegation(ctx sdk.Context, red types.Redelegation
 	store.Set(redKey, bz)
 	store.Set(GetREDByValSrcIndexKey(red.DelegatorAddr, red.ValidatorSrcAddr, red.ValidatorDstAddr, k.cdc), redKey)
 	store.Set(GetREDByValDstIndexKey(red.DelegatorAddr, red.ValidatorSrcAddr, red.ValidatorDstAddr, k.cdc), redKey)
+}
+
+// remove a redelegation object and associated index
+func (k PrivlegedKeeper) RemoveRedelegation(ctx sdk.Context, red types.Redelegation) {
+	store := ctx.KVStore(k.storeKey)
+	redKey := GetREDKey(red.DelegatorAddr, red.ValidatorSrcAddr, red.ValidatorDstAddr, k.cdc)
+	store.Delete(redKey)
+	store.Delete(GetREDByValSrcIndexKey(red.DelegatorAddr, red.ValidatorSrcAddr, red.ValidatorDstAddr, k.cdc))
+	store.Delete(GetREDByValDstIndexKey(red.DelegatorAddr, red.ValidatorSrcAddr, red.ValidatorDstAddr, k.cdc))
 }